feat: RefreshToken 엔티티 정의(#3)
