--- src/main/java/com/specialwarriors/conal/common/jwt/domain/RefreshToken.java
@@ -1,21 +1,0 @@
-package com.specialwarriors.conal.common.jwt.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Long userId;
-
-    @Column(nullable = false)
-    private String refreshToken;
-}
